Tests/LibWeb/Text/expected/canvas/webgl-draw-does-not-invalidate-display-list.txt

[webgl] needsDisplayListRecord after getContext: false
[webgl] needsDisplayListRecord before draw: false
[webgl] needsRepaint after draw: true
[webgl] needsDisplayListRecord after draw: true
[webgl2] needsDisplayListRecord after getContext: false
[webgl2] needsDisplayListRecord before draw: false
[webgl2] needsRepaint after draw: true
[webgl2] needsDisplayListRecord after draw: true
